port OverlayTriggerSpec.tsx to rtl
